# be/src/comments/dto/comment-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { UserSummaryDto } from '../../users/dto/user-summary.dto';

// 댓글 항목 응답
export class CommentResponseDto {
  @ApiProperty({ format: 'uuid' })
  id: string;

  @ApiProperty({ format: 'uuid' })
  postId: string;

  @ApiProperty({ type: () => UserSummaryDto })
  author: UserSummaryDto;

  @ApiProperty({ example: '멋진 사진이네요!' })
  content: string;

  @ApiProperty()
  createdAt: Date;

  @ApiProperty()
  updatedAt: Date;
}